+++ cameraview/src/test/java/com/otaliastudios/cameraview/size/AspectRatioTest.java
@@ -1,8 +1,4 @@
-package com.otaliastudios.cameraview;
-
-
-import com.otaliastudios.cameraview.size.AspectRatio;
-import com.otaliastudios.cameraview.size.Size;
+package com.otaliastudios.cameraview.size;
 
 import org.junit.Test;
